Create MenuItem and MenuItemImage components
