+++ client/src/Menu.js
@@ -1,16 +1,4 @@
-import './App.css'
-
-function MenuItem({ name, image, description }) {
-  return (
-    <div className="menu-item">
-      <h3 className="menu-item-name">{name}</h3>
-      <div className="menu-img-container">
-        <img className="menu-img" src={image.path} alt={image.alt} />
-        <p>{description}</p>
-      </div>
-    </div>
-  )
-}
+import MenuItem from "./MenuItem"
 
 function Menu() {
   return (
